Correct installation ini file path (/etc/ohm/plugins.d).
The chain-effects of the wrong path are astonishing...
Rule engine cannot find the config file so it will not
preinitalize itself. It expects someone to configure it
via the exported setup mechanism (which is not done any
more by the resolver plugin). Since no initial configuration
exists predicates are not discovered during initialization.

At the first rule evaluation attempt rule_engine tries to discover
the rule predicates. This involves looking up and invoking the prolog
predicate exported/1. Here is where the real fun begins: when you look
up a prolog predicate with PL_pred or PL_predicate you always get back
a (formally) valid predicate handle. If the predicate does not exist
at the time you look it up, a new one will be allocated and hashed in
for you. The new predicate will have no clauses, etc. When you try to
evaluate such a predicate the interpreter plain simply crashes.

As a workaround for now I will make rule evaluation fail if the plugin
has not been successfully setup either during initialization or later
via the exported setup.

However this does not fully solve the problem as such. ATM I do not know
how to reliably tell about a predicate whether it is defined or not and
we'd really need one. Maybe trying to look up the predicates clauses and
check that they are not empty could be a close approximation. Needs
further investigation....