Prolog helper/glue code lifted over from policy-settings to libprolog.pl.
The prolog glue code implements part of the functionality provided by this
library. It relies heavily on foreign predicates implemented in this library
and vice versa, this library uses the predicates of the glue code among other
things for load-time error detection, predicate tracing, and predicate
discovery. Hence the glue code should (always have) be(en) part of this
library.

Rulesets do not need to include and load a copy of it. It gets automatically
loaded by the library initialization routines.