Cache and reuse predicate lookup results internally.
With the original prolog_{rules,*call,statistics,free_predicates}
semantics it was not possible to discover predicates more than once
and get reasonable statistics. Statistics were collected on a per
prolog_predicate_t array basis. This made it difficult to query
statistics without running the predicate evaluation. To work around
this limitation without breaking the API we now cache the result
of predicate lookup internally. As a side-effect of this freeing
the predicates happens now also internally during library cleanup.
Hence prolog_free_predicates is a no-op when called from outside
the library.