Do not set SWI_HOME_DIR and especially not using putenv(3) !!!
Depending on the version of (g)libc, putenv might put the supplied
string directly to the environment. Interesting things happen if
such an implementation of putenv is called from a dlopened shared
object and that object is later dlclosed. A typical symptom is that
functions examining the environment (eg. localtime and strftime)
start crashing.

Ouch, this one was pretty nasty to hunt down...